BCB-924 move abstract class into component and add retry
